Refactor add state objects
