+++ core/date-time/src/commonMain/kotlin/xyz/ksharma/krail/core/datetime/DateTimePickerInfo.kt
@@ -24,7 +24,7 @@ fun rememberCurrentDateTime(): DateTimePickerInfo {
 
     // Formatting the date
     val formattedDate = remember(currentDateTime.date) {
-        formatDate(currentDateTime.date)
+        toReadableDate(currentDateTime.date)
     }
 
     return DateTimePickerInfo(
@@ -32,21 +32,6 @@ fun rememberCurrentDateTime(): DateTimePickerInfo {
         hour = currentDateTime.hour,
         minute = currentDateTime.minute,
     )
-}
-
-fun formatDate(date: LocalDate): String {
-    val today = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).date
-    val tomorrow = today.plus(1, DateTimeUnit.DAY)
-
-    return when (date) {
-        today -> "Today"
-        tomorrow -> "Tomorrow"
-        else -> {
-            val dayOfWeek = date.dayOfWeek.name.substring(0, 3) // Short day name (e.g., Mon, Tue)
-            val month = date.month.name.substring(0, 3) // Short month name (e.g., Jan, Feb)
-            "$dayOfWeek ${date.dayOfMonth} ${month.replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }}"
-        }
-    }
 }
 
 fun incrementDateByOneDay(date: LocalDate): LocalDate {
@@ -57,9 +42,30 @@ fun decrementDateByOneDay(date: LocalDate): LocalDate {
     return date.plus(-1, DateTimeUnit.DAY)
 }
 
-fun formatTime(hour: Int, minute: Int): String {
+/**
+ * Formats the time in 12-hour format (e.g., 1:30 PM)
+ */
+fun to12HourTimeString(hour: Int, minute: Int): String {
     val displayHour = if (hour == 0 || hour == 12) 12 else hour % 12
     val amPm = if (hour < 12) "AM" else "PM"
     val formattedMinute = if (minute < 10) "0$minute" else minute.toString()
     return "$displayHour:$formattedMinute $amPm"
+}
+
+/**
+ * Formats the date in a human-readable format (e.g., Today, Tomorrow, Mon 1 Jan)
+ */
+fun toReadableDate(date: LocalDate): String {
+    val today = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).date
+    val tomorrow = today.plus(1, DateTimeUnit.DAY)
+
+    return when (date) {
+        today -> "Today"
+        tomorrow -> "Tomorrow"
+        else -> {
+            val dayOfWeek = date.dayOfWeek.name.substring(0, 3) // Short day name (e.g., Mon, Tue)
+            val month = date.month.name.substring(0, 3) // Short month name (e.g., Jan, Feb)
+            "$dayOfWeek, ${date.dayOfMonth} ${month.replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }}"
+        }
+    }
 }
