+++ feature/trip-planner/state/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/state/datetimeselector/JourneyTimeOptions.kt
@@ -1,0 +1,6 @@
+package xyz.ksharma.krail.trip.planner.ui.state.datetimeselector
+
+enum class JourneyTimeOptions(val text: String) {
+    LEAVE("Leave"),
+    ARRIVE("Arrive")
+}
